// packages/generator/src/emitter/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type UniqueIndex = ReadonlyDeep<{
    name: string;
    fields: string[];
  }>;
  export type PrimaryKey = ReadonlyDeep<{
    name: string | null;
    fields: string[];
  }>;
  export type Model = ReadonlyDeep<{
    name: string;
    dbName: string | null;
    fields: ModelField[];
    uniqueFields: string[][];
    uniqueIndexes: UniqueIndex[];
    // documentation?: string;
    primaryKey: PrimaryKey | null;
    // isGenerated?: boolean;
    // additional props:
    typeName: string;
    docs: string | undefined;
    plural: string | undefined;
    isOmitted: { output: boolean };
  }>;
  export type FieldKind = "scalar" | "object" | "enum" | "unsupported";
  export type FieldNamespace = "model" | "prisma";
  export type FieldLocation =
    | "scalar"
    | "inputObjectTypes"
    | "outputObjectTypes"
    | "enumTypes"
    | "fieldRefTypes";
  // Field
  export type ModelField = ReadonlyDeep<{
    // kind: FieldKind;
    name: string;
    isRequired: boolean;
    isList: boolean;
    isUnique: boolean;
    isId: boolean;
    isReadOnly: boolean;
    isGenerated?: boolean;
    isUpdatedAt?: boolean;
    // type: string;
    dbNames?: string[] | null;
    hasDefaultValue: boolean;
    default?: FieldDefault | FieldDefaultScalar | FieldDefaultScalar[];
    relationFromFields?: string[];
    relationToFields?: string[];
    relationOnDelete?: string;
    relationName?: string;
    // documentation?: string;
    // [key: string]: any;
    // additional props:
    type: string;
    location: FieldLocation;
    typeFieldAlias?: string;
    typeGraphQLType: string;
    fieldTSType: string;
    docs: string | undefined;
    isOmitted: { input: boolean | InputOmitSetting[]; output: boolean };
  }>;
// ...
  export type ModelMapping = ReadonlyDeep<{
    // model: string;
    // plural: string;
    // findUnique?: string | null;
    // findUniqueOrThrow?: string | null;
    // findFirst?: string | null;
    // findFirstOrThrow?: string | null;
    // findMany?: string | null;
    // createOne?: string | null;
    // createMany?: string | null;
    // createManyAndReturn?: string | null;
    // updateOne?: string | null;
    // updateMany?: string | null;
    // upsertOne?: string | null;
    // deleteOne?: string | null;
    // deleteMany?: string | null;
    // aggregate?: string | null;
    // groupBy?: string | null;
    // count?: string | null;
    // findRaw?: string | null;
    // aggregateRaw?: string | null;

    // additional props
    modelName: string;
    actions: Action[];
    collectionName: string;
    resolverName: string;
    modelTypeName: string;
  }>;
// ...
  export type Action = ReadonlyDeep<{
    name: string;
    fieldName: string;
    kind: ModelAction;
    operation: "Query" | "Mutation";
    prismaMethod: string;
    method: OutputSchemaField;
    argsTypeName: string | undefined;
    outputTypeName: string;
    actionResolverName: string;
    returnTSType: string;
    typeGraphQLType: string;
  }>;
  // additional type
  export type RelationModel = ReadonlyDeep<{
    model: Model;
    outputType: OutputType;
    relationFields: RelationField[];
    resolverName: string;
  }>;
  // additional type
  export type RelationField = ModelField &
    ReadonlyDeep<{
      outputTypeField: OutputSchemaField;
      argsTypeName: string | undefined;
    }>;
}